Put ASWebAuthenticationSessionUIImplementation in its own file
